requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
httpx==0.25.1
